+++ NEWS.md
@@ -1,6 +1,11 @@
 News
 =====
 
+# glmmSeq v0.5.0
+###### 15/8/2022
+* Add option to `glmmSeq()` to use `glmmTMB` package for fitting negative
+binomial GLMM (or other GLM family) models.
+
 # glmmSeq v0.4.0
 ###### 10/8/2022
 * Further speed enhancements to `lmmSeq()` using `lme4::modular` code. Speed